+++ hata/discord/oauth2/__init__.py
@@ -3,6 +3,7 @@ from .connection import *
 from .helpers import *
 from .oauth2_access import *
 from .oauth2_user import *
+from .preinstanced import *
 
 __all__ = (
     *achievement.__all__,
@@ -10,4 +11,5 @@ __all__ = (
     *helpers.__all__,
     *oauth2_access.__all__,
     *oauth2_user.__all__,
+    *preinstanced.__all__,
 )